// src/pokemon/models/pokemon.ts
export interface PokemonBase {
  name: string;
  url: string;
}

export interface Pokemon extends PokemonBase {
  id: number;
  height: number;
  weight: number;
  sprites: Sprites;
  isFavorite: boolean;
}
// ...
export interface Sprites {
  back_default: string;
  back_female?: any;
  back_shiny: string;
  back_shiny_female?: any;
  front_default: string;
  front_female?: any;
  front_shiny: string;
  front_shiny_female?: any;
  other: Other | null;
  versions: Versions | null;
}

interface Type {
  slot: number;
  type: Ability;
}

interface Stat {
  base_stat: number;
  effort: number;
  stat: Ability;
}

interface Versions {
  'generation-i': GenerationI;
  'generation-ii': GenerationII;
  'generation-iii': GenerationIII;
  'generation-iv': GenerationIV;
  'generation-v': GenerationV;
  'generation-vi': GenerationVI;
  'generation-vii': GenerationVII;
  'generation-viii': GenerationVIII;
}
// ...
interface Other {
  dream_world: Dreamworld;
  home: Home;
  'official-artwork': OfficialArtwork;
}